[2218] Change back size of DomainTreeNodeChain's level counter to size_t

+++ src/lib/datasrc/memory/domaintree.h
@@ -770,7 +770,7 @@ public:
     /// chain, 0 will be returned.
     ///
     /// \exception None
-    uint32_t getLevelCount() const { return (level_count_); }
+    size_t getLevelCount() const { return (level_count_); }
 
     /// \brief return the absolute name for the node which this
     /// \c DomainTreeNodeChain currently refers to.
@@ -788,7 +788,7 @@ public:
 
         const DomainTreeNode<T>* top_node = top();
         isc::dns::Name absolute_name = top_node->getName();
-        uint32_t level = level_count_ - 1;
+        size_t level = level_count_ - 1;
         while (level > 0) {
             top_node = nodes_[level - 1];
             absolute_name = absolute_name.concatenate(top_node->getName());
@@ -848,7 +848,7 @@ private:
     // it's also equal to the possible maximum level.
     const static int RBT_MAX_LEVEL = isc::dns::Name::MAX_LABELS;
 
-    uint32_t level_count_;
+    size_t level_count_;
     const DomainTreeNode<T>* nodes_[RBT_MAX_LEVEL];
     const DomainTreeNode<T>* last_compared_;
     isc::dns::NameComparisonResult last_comparison_;
